Do the same for ESP32-S2/S3...

+++ esp-hal/src/clock/clocks_ll/esp32s2.rs
@@ -38,8 +38,6 @@ pub(crate) fn set_cpu_clock(cpu_clock_speed: CpuClock) {
         });
 
         let value = (((80 * MHZ) >> 12) & UINT16_MAX) | ((((80 * MHZ) >> 12) & UINT16_MAX) << 16);
-        rtc_cntl
-            .store5()
-            .modify(|_, w| w.scratch5().bits(value as u32));
+        rtc_cntl.store5().modify(|_, w| w.scratch5().bits(value));
     }
 }
